fixes formatDuration for durations > 1 minute
Format string will respond with an empty string when its given a duration less than a minute.
While this scenario is pretty rare a user did run into this issue with the daily command which calls formatDuration.

see: (https://cdn.imgchest.com/files/vmy2pc5lk7j.png) and (https://cdn.imgchest.com/files/rj7kzcx2p7m.png)
for user who ran into the issue's post in the help-and-support channel and the support response.

I'm not too familiar with JS or this project as a whole so I'm not sure if this fix will cause issues with anything else that calls formatDuration. Additionally this was prototyped in my browsers JS console so it hasn't been tested in a fully fledged dev environment for the project.