split up the monolithic +cl clues command into seperate commands: +cl clues shared +cl clues beginner +cl clues easy +cl clues medium +cl clues hard +cl clues elite +cl clues master +cl clues rare (for 3a and gilded items)
the command
+cl clues all
was added to do what +cl clues used to do and prints out a cl of all the clue items

I also added aliases for the commands so the "clues" part isnt needed (ex. +cl hard) and if the "clues" part is non plural (ex. +cl clue hard)
and an alias to shorten medium the various medium teir commands to accept med