fix and clean up +cl clue (#146)
* +cl clue clean up and reorganization

no items should be outside the bounds of the image and items are now in the order of clue teir:
	shared, beginner, easy, medium, hard, elite, master

* Revert "+cl clue clean up and reorganization" forgot to lint

This reverts commit 7901904fa694503adfb688ba0e8c3ec2c04a13fb.

* +cl clue clean up and reorganization

now properly linted aswell

no items should be outside the bounds of the image and items are now in the order of clue teir:
            shared, beginner, easy, medium, hard, elite, master

* split up the monolithic +cl clues command into seperate commands:
+cl clues shared
+cl clues beginner
+cl clues easy
+cl clues medium
+cl clues hard
+cl clues elite
+cl clues master
+cl clues rare (for 3a and gilded items)

the command
+cl clues all
was added to do what +cl clues used to do and prints out a cl of all the clue items

I also added aliases for the commands so the "clues" part isnt needed (ex. +cl hard) and if the "clues" part is non plural (ex. +cl clue hard)
and an alias to shorten medium the various medium teir commands to accept med

* forgot to lint.

* forgot to add cluesBegginer to cluesAll

also swapped cluesAll from using Object.assign to using a object spread

* and as per usual forgot to lint

* added the clueBeginner in the worng spot

* removed //usageDelim: ' '

I only commented it out for testing at first but never got back around to fully removing it