Bank Sale For Running Sell Command On Array
This is a rough draft for a command that will take an array of arrays like so:
```+bankSale [["Rune battleaxe",5], ["Feather",""],["Bass",100000]]```
It will then process each underlying array by passing the Item name and quantity into the functions I mirrored from the Sell command. 
I intended for it to only be able to process a set number of items by setting the Limit variable. This could be dynamically changed based on the user's patron status in order to allow for more lines to be processed in this command.
If the user does not specify a quantity for the item being sold, it will pull the maximum amount that user has in their bank, and set that as the quantity.