dryCalc.js
This will take two numbers, one being the drop rate for an item, and the other being the current number of kills that the person has earned towards the item. This will track how "dry" the person is by outputting the likeliness of them getting the drop at the boss.