Add warrior helm to buyables
I scaled the price from 78,000 gp to 780,000 gp based on the scaling of Helm of Neitiznot from 50,000 gp to 500,000 gp (10x original price).

I used a 60 quest point requirement to reflect the fact that The Fremennik Trials is a prerequisite to The Fremennik Isles and thus this helm should require fewer quest points than the Helm of Neitiznot. The reason for adding the helm is that it's useful for CoX melee gear setups as well as buyable in the real game.