Add feed command for hammy (#1930)
* A slightly buffed Hammy.
This one has a confirm option + --cf/--yes flag.
It also has a 1 in 10 chance to save the item.
With a rare 1 in 100 chance to double the item.

And of course, more fun Hammy hijinx! Hammy may or may not have a slight gambling addiction...

* A slightly buffed Hammy.
This one has a confirm option + --cf/--yes flag.
It also has a 1 in 10 chance to save the item.
With a rare 1 in 100 chance to double the item.

And of course, more fun Hammy hijinx! Hammy may or may not have a slight gambling addiction... ESLint

* A slightly buffed Hammy.
This one has a confirm option + --cf/--yes flag.
It also has a 1 in 10 chance to save the item.
With a rare 1 in 100 chance to double the item.

And of course, more fun Hammy hijinx! Hammy may or may not have a slight gambling addiction... ESLint2

* Fix review