Dinh's now prevents cannon/barrage
This is consistent with in game.

"Furthermore, players wielding the bulwark will not be able to cast any combat spells; players attempting to do so will receive the message "Your bulwark gets in the way". When in defensive mode, the dwarf multicannon will not fire, but it will still rotate, with a filterable message stating "Your bulwark shuns your cowardliness; stopping your cannon from firing.""