Tome of water/fire rune reductions
Both tome of water and fire should reduce runes for skilling, without consuming charges.

Water tome does not consume charges for ice spells, which are the only combat spells the bot is using water runes for. If we add rune costs in the future for combat spells, this needs to be revisited.

Added both tomes to skilling rune reductions, regardless of charge.