Collection Log updates
Remove the individual dks kings cl, add aliases for the indvidiual dks kings under the group cl.

Added the individual dks kings kcs to the top of this cl.

Changed a few cl's with multiple --all items to be working and not just referencing the first in the list thanks to Lux