Daeyalt Essence
Adds the mining of daeyalt shards to gather daeyalt essence (auto banks the essence instead of the shards as that's their only purpose)
Adds a boolean optional command on the runecrafting command to use daeyalt essence at a 50% increased xp rate. (doesn't work on blood/soul runes)
If the boolean is not selected it defaults to false.