CL Custom unobtainable black mask fix
Currently the black mask is on the cl custom but unobtainable as it's autoconverted from the (10) version on the drops.

I also tried checking in this command still works if an old user had a black mask (10) in their bank but it wouldn't work for me. So deleting this command seems the best route.