Allow fish barrel to extend trips
This is a redo of PR #2659

Allows the fish barrel/fish sack barrel to extend any trips under the fish command by 9 minutes.
Adds a message showing that the boost is working.

Can't see any reason why this can't be on the osb branch. Believe the fish barrel extend will have to be deleted off the bso branch if this gets merged to prevent duplication.