Change Igne Wiki back to boss, rather than tame
For some reason in this commit -https://github.com/oldschoolgg/oldschoolbot/commit/aaa4b3362dea0b87d797f2a379b605343435ec5d

The igne wiki page for the boss was edited and made to be a copy of the page about the igne tame. This reverts it back to how it was prior.